+++ test/test_helper.rb
@@ -13,7 +13,7 @@ if $VERBOSE
     require 'warning'
   rescue LoadError
   else
-    Warning.ignore(%r{lib/wikicloth/extensions/|lib/pdf/reader/font.rb|lib/maruku/|lib/creole/parser.rb|lib/commonmarker})
+    Warning.ignore(%r{lib/pdf/reader/font.rb|lib/creole/parser.rb|lib/commonmarker})
   end
 end